extract fullscreen player into a standalone widget
